requirements.txt
Flask==0.12
flask-socketio
flask-mail==0.9.1
Flask-PageDown==0.2.2
Flask-Admin==1.4.2
Flask-Moment==0.5.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.10
WTForms==1.0.5
sqlalchemy
flask_sqlalchemy
PyMySQL==0.7.9
oauth2client>=2.0.0
Markdown==2.6.7
Flask-PyMongo==0.5.1
PyMongo==3.4.0
xmljson
gevent
gunicorn
requests
celery
Pillow==4.3.0
flask-dropzone==1.4